locked down office editing to mainly site admins only (especially deletes)

+++ lib/mpnetwork_web/controllers/office_controller.ex
@@ -1,12 +1,12 @@
 defmodule MpnetworkWeb.OfficeController do
   use MpnetworkWeb, :controller
 
-  alias Mpnetwork.Realtor
-  alias Mpnetwork.Realtor.Office
+  alias Mpnetwork.{Realtor, Permissions}
+  alias Realtor.Office
 
   def index(conn, _params) do
     offices = Realtor.list_offices()
-    render(conn, "index.html", offices: offices)
+    render(conn, "index.html", offices: offices, show_controls: Permissions.office_admin_or_site_admin?(current_user(conn)))
   end
 
   def new(conn, _params) do
@@ -15,13 +15,17 @@ defmodule MpnetworkWeb.OfficeController do
   end
 
   def create(conn, %{"office" => office_params}) do
-    case Realtor.create_office(office_params) do
-      {:ok, office} ->
-        conn
-        |> put_flash(:info, "Office created successfully.")
-        |> redirect(to: office_path(conn, :show, office))
-      {:error, %Ecto.Changeset{} = changeset} ->
-        render(conn, "new.html", changeset: changeset)
+    if Permissions.office_admin_or_site_admin?(current_user(conn)) do
+      case Realtor.create_office(office_params) do
+        {:ok, office} ->
+          conn
+          |> put_flash(:info, "Office created successfully.")
+          |> redirect(to: office_path(conn, :show, office))
+        {:error, %Ecto.Changeset{} = changeset} ->
+          render(conn, "new.html", changeset: changeset)
+      end
+    else
+      send_resp(conn, 405, "Not allowed")
     end
   end
 
@@ -31,30 +35,42 @@ defmodule MpnetworkWeb.OfficeController do
   end
 
   def edit(conn, %{"id" => id}) do
-    office = Realtor.get_office!(id)
-    changeset = Realtor.change_office(office)
-    render(conn, "edit.html", office: office, changeset: changeset)
+    if Permissions.office_admin_or_site_admin?(current_user(conn)) do
+      office = Realtor.get_office!(id)
+      changeset = Realtor.change_office(office)
+      render(conn, "edit.html", office: office, changeset: changeset)
+    else
+      send_resp(conn, 405, "Not allowed")
+    end
   end
 
   def update(conn, %{"id" => id, "office" => office_params}) do
-    office = Realtor.get_office!(id)
+    if Permissions.office_admin_or_site_admin?(current_user(conn)) do
+      office = Realtor.get_office!(id)
 
-    case Realtor.update_office(office, office_params) do
-      {:ok, office} ->
-        conn
-        |> put_flash(:info, "Office updated successfully.")
-        |> redirect(to: office_path(conn, :show, office))
-      {:error, %Ecto.Changeset{} = changeset} ->
-        render(conn, "edit.html", office: office, changeset: changeset)
+      case Realtor.update_office(office, office_params) do
+        {:ok, office} ->
+          conn
+          |> put_flash(:info, "Office updated successfully.")
+          |> redirect(to: office_path(conn, :show, office))
+        {:error, %Ecto.Changeset{} = changeset} ->
+          render(conn, "edit.html", office: office, changeset: changeset)
+      end
+    else
+      send_resp(conn, 405, "Not allowed")
     end
   end
 
   def delete(conn, %{"id" => id}) do
-    office = Realtor.get_office!(id)
-    {:ok, _office} = Realtor.delete_office(office)
+    if Permissions.site_admin?(current_user(conn)) do
+      office = Realtor.get_office!(id)
+      {:ok, _office} = Realtor.delete_office(office)
 
-    conn
-    |> put_flash(:info, "Office deleted successfully.")
-    |> redirect(to: office_path(conn, :index))
+      conn
+      |> put_flash(:info, "Office deleted successfully.")
+      |> redirect(to: office_path(conn, :index))
+    else
+      send_resp(conn, 405, "Not allowed, need to be a site administrator to delete offices")
+    end
   end
 end
